refactor: API 엔드포인트 경로 재설정 및 구조 개선
