+++ src/main/java/com/pado/domain/study/controller/StudyController.java
@@ -49,16 +49,6 @@ public class StudyController {
         return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
-    @Operation(summary = "내 스터디 목록 조회", description = "현재 로그인한 사용자가 참여하고 있는 스터디 목록을 조회합니다.")
-    @ApiResponse(responseCode = "200", description = "조회 성공")
-    @GetMapping("/me")
-    public ResponseEntity<List<MyStudyResponseDto>> getMyStudies(
-        @Parameter(hidden = true) @CurrentUser User user
-    ) {
-        List<MyStudyResponseDto> myStudies = studyService.findMyStudies(user.getId());
-        return ResponseEntity.ok(myStudies);
-    }
-
     @SecurityRequirements({})
     @Operation(summary = "스터디 목록 조회 및 검색",
         description = """
